Optionally filter remarks by those added by the current git branch
This enables a really handy way of keeping track of development todos on
a branch: just write XXX remarks in your work-in-progress commits,
remove them as you address those items, and run `xxx -b` to see what
remains after each commit.

Currently this only works if run from within a git working directory.
It would be possible to make work from _outside_ a set of git working
dirs, but that's not a feature I currently need.